delete ss_jmvae and add ss_mvae
